Remove bad animation ranges with discard range events (#12)
We got a mechanism in place to discard specific frames/ranges of an animation and exclude it from the motion matching database. We have several sections in our animations where the arms are over the head, where the person suddenly crouches for a bit or other poses that do not match the other locomotion data. I went through the animation database and discarded all of these sections in order to improve the results of the synthesized animations.

Signed-off-by: Benjamin Jillich <43751992+amzn-jillich@users.noreply.github.com>