Fix issue with debug console ignoring some keys
This problem was reported for Server only, the Enter/Backspace keys were
being ignored in the ImGui Debug Console.
This wasn't an issue if the Server had loaded a map.
The problem was with XConsole explicitly setting a bool in dedicated
server mode.  This caused text input to be processed by XConsole code
and not passed further along to DebugConsole where it should have been
handling it via ImGui.

Signed-off-by: amzn-phist <52085794+amzn-phist@users.noreply.github.com>