Fixed test_WindowsAndMac_FilesMarkedSkip_FilesAreSkipped to pass again.
Updated the test to verify the assets are actually skipped and not just missing, by having it first run without the skip command and verify they are in the first run. Also updated logging to print out sorted lists, to make it easier to debug failures on Jenkins in the future.

Signed-off-by: AMZN-stankowi <4838196+AMZN-stankowi@users.noreply.github.com>