Automated testing create_backup and restore_backup respects file flags.
Example:
If the file that is to be backed-up is readonly, when restored that file will be readonly.
If the file that is to be backed-up is not readonly, when restored that file will be not readonly.

Fixes an issue that would mark physics setreg files to readonly after running AutomatedTesting when restoring from the backup.

Signed-off-by: amzn-sean <75276488+amzn-sean@users.noreply.github.com>