Motion Matching: Motion extraction sometimes outputs zero movement and results in stuttering (#22)
The problem was that after the motion matching algorithm determined a better matching frame and we started a cross-fade, at that given frame the motion extraction delta was outputting a zero vector, resulting in a little stutters at the mm update frequency. This happened because the exrtaction calculation was done before the motion instance time values were updated and also the previous time values weren't set correctly when switching the frame/animation.

Signed-off-by: Benjamin Jillich <43751992+amzn-jillich@users.noreply.github.com>