Reduce number of includes in AssetDataStream.h
PIMPL away members of AssetDataStream that needed the heavy include files.

Move ReportType out of ReportData to allow forward declarations.

When destructor is defaulted/defined in the header file,
all of the class member destructors must all be instantiable in the
header as well. By moving Scheduler destructor to cpp file, we are spared some inclusions.

Signed-off-by: nemerle <96597+nemerle@users.noreply.github.com>