Fixed the translation error messages for GraphCanvas
It was unconditionally logging an AZ_Warning and AZ_Error if a
translation entry with the same key was added to the database.
This didn't account for the same key and value being added to the
database, so loading the same translation file twice would result in
error messages for each entry.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>