[Linux] Prefer higher versioned clang compilers to lower versions (#6824)
If the user does not specify which compiler to use via cache variables in
the CMake invocation, we perform a search for which version to use.
Previously this was done using a glob search on some pre-defined paths, and
then sorting the results using `list(SORT ... COMPARE NATURAL)`. Sorting in
this manner results in lower versions being moved to the head of the list.
The first result in the list was then used. Consequently, if the user has
`clang-11` and `clang-12` installed, `clang-11` was chosen.

This reverses the sort order so that the highest installed version is
chosen.

Signed-off-by: Chris Burel <burelc@amazon.com>
Signed-off-by: santorac <55155825+santorac@users.noreply.github.com>