Updating mock calls in unit_test_manifest.py to fix test
The manifest.py `get_all_templates` method was modified to add a call to
`get_all_gems` and `get_gem_templates`.
Those calls were not mocked, so they tried to load an o3de_manifest.json
file that was on user's machine, which failed on the CI node.

Also added a mock call for `load_o3de_manifest`, that validates that it
is not called to catch this issue during Automated Review.

Signed-off-by: lumberyard-employee-dm <56135373+lumberyard-employee-dm@users.noreply.github.com>