Irradiance Color taken from Base Color by default
This change is related to feature request #6808.

The irradiance color can now be selected via an irradiance.irradianceColorSource
enum to be automatically derived from the base color (the new default), or set
to a user-provided irradiance.manualColor value (as it was previously).
When using the base color as irradianceColorSource together with a base color
texture, the average texture color is used as a basis for the irradiance color
(currently only supported for StreamingImageAsset textures).

NOTE: There is mild breakage but no loss of data when auto-converting old
.material files:
The 'toVersion' rename operations translate the old 'irradiance.color' to the
current 'irradiance.manualColor', but the new irradianceColorSource will be set
to its default 'BaseColor' value, i.e. the old irradiance color is saved but not
used. The user needs to explicitly set irradiance.irradianceColorSource to
'Manual' to recover the old behaviour. (A 'setValue' type of versionUpdate
operation that sets the irradianceColorSource to 'Manual' could maybe automate
this?)

Signed-off-by: Roald Frederickx <roald.frederickx@huawei.com>