Generate AllProperties material files.
The MaterialTypeBuilder will new generate, in addition to the material type asset, a <MaterialType>_AllProperties.material file that specifies the defailt value for all available properties. It is stored in the cache next to the .azmaterialtype file. This can be convenient as reference documentation, a reference, or template, especially useful for TA's working on scripted data pipelines. It is enabled by default but can be disabled by setting "/O3DE/Atom/RPI/MaterialTypeBuilder/CreateAllPropertiesMaterial" in the settings registry.

I also added a missing error report in JsonMaterialPropertyValueSerializer, and updated a property value warning to include the property name, both of which were helpful in address issues I encountered along the way.

Signed-off-by: santorac <55155825+santorac@users.noreply.github.com>