Add break on assert setting with debugger attached
This commit adds a new "O3DE/Core/assert_auto_break" settings register
key which defaults currently to `false`. When enabled, trace asserts
will check if a debugger is attached and emit a debugger interrupt if
so.

The idea is that after introduction, developers will be *strongly*
encouraged to enable this flag so that active asserts currently are
fixed. After a period of time, this setting will default to true so that
asserts that trigger are more visible across the community.

Signed-off-by: Jeremy Ong <jcong@amazon.com>